+++ src/editor-mathfield/mathfield-private.ts
@@ -1481,6 +1481,13 @@ export class MathfieldPrivate implements GlobalContext, Mathfield {
       })
     );
 
+    this.host?.dispatchEvent(
+      new UIEvent('focusin', {
+        bubbles: true, // unlike 'focus', focusin does bubble
+        composed: true,
+      })
+    );
+
     // Save the current value.
     // It will be compared in `onBlur()` to see if the
     // `change` event needs to be dispatched. This
@@ -1498,8 +1505,12 @@ export class MathfieldPrivate implements GlobalContext, Mathfield {
     this.ariaLiveText!.textContent = '';
 
     complete(this, 'accept');
-    if (this.model.getValue() !== this.valueOnFocus)
-      this.executeCommand('commit');
+
+    if (this.model.getValue() !== this.valueOnFocus) {
+      this.host?.dispatchEvent(
+        new Event('change', { bubbles: true, composed: true })
+      );
+    }
 
     if (
       /onfocus|manual/.test(this.options.virtualKeyboardMode) &&
@@ -1512,6 +1523,13 @@ export class MathfieldPrivate implements GlobalContext, Mathfield {
     this.host?.dispatchEvent(
       new Event('blur', {
         bubbles: false, // DOM 'focus' and 'blur' don't bubble
+        composed: true,
+      })
+    );
+
+    this.host?.dispatchEvent(
+      new UIEvent('focusout', {
+        bubbles: true, // unlike 'blur', focusout does bubble
         composed: true,
       })
     );