+++ src/core/atom-class.ts
@@ -1372,8 +1372,7 @@ function renderStyleRun(
   // In most cases we want to display selection,
   // except if the `atomIdsSettings.groupNumbers` flag is set which is used for
   // read aloud.
-  const displaySelection =
-    !context.atomIdsSettings || !context.atomIdsSettings.groupNumbers;
+  const displaySelection = !context.atomIdsSettings?.groupNumbers;
 
   let boxes: Box[] = [];
   let newList = options?.newList ?? false;