+++ src/editor-mathfield/commands.ts
@@ -67,20 +67,13 @@ registerCommand({
     mathfield.insert('.');
     return true;
   },
+  // A 'commit' command is used to simulate pressing the return/enter key,
+  // e.g. when using a virtual keyboard
   commit: (mathfield: MathfieldPrivate) => {
     if (contentWillChange(mathfield.model, { inputType: 'insertLineBreak' })) {
-      // No matching keybinding: trigger a commit
-
-      if (mathfield.host) {
-        mathfield.host.dispatchEvent(
-          new Event('change', {
-            bubbles: true,
-            composed: true,
-          })
-        );
-      }
-
-      // Dispatch an 'input' event matching the behavior of `<textarea>`
+      mathfield.host?.dispatchEvent(
+        new Event('change', { bubbles: true, composed: true })
+      );
       contentDidChange(mathfield.model, { inputType: 'insertLineBreak' });
     }
     return true;