+++ src/editor-mathfield/keyboard-input.ts
@@ -97,7 +97,7 @@ export function onKeystroke(
   // Ignore the key if Command or Control is pressed (it may be a keybinding,
   // see 4.3)
   if (!mathfield.options.readOnly) {
-    if (mathfield.mode === 'math' && !evt.ctrlKey && !evt.metaKey) {
+    if (mathfield.mode === 'math') {
       if (keystroke === '[Backspace]') {
         // Special case for backspace to correctly handle undoing
         mathfield.inlineShortcutBuffer.pop();
@@ -105,7 +105,6 @@ export function onKeystroke(
       } else if (!mightProducePrintableCharacter(evt)) {
         // It was a non-alpha character (PageUp, End, etc...)
         mathfield.flushInlineShortcutBuffer();
-        mathfield.snapshot();
       } else {
         const c = eventToChar(evt);
 
@@ -213,10 +212,7 @@ export function onKeystroke(
 
         if (mathfield.host) {
           result = !mathfield.host.dispatchEvent(
-            new Event('change', {
-              bubbles: true,
-              composed: true,
-            })
+            new Event('change', { bubbles: true, composed: true })
           );
         }
 
@@ -555,7 +551,7 @@ function getLeftSiblings(mf: MathfieldPrivate): Atom[] {
   let atom = model.at(Math.min(model.position, model.anchor));
   while (atom.type !== 'first') {
     result.push(atom);
-    atom = atom.leftSibling;
+    atom = atom.leftSibling!;
   }
 
   return result;