+++ src/editor-model/delete.ts
@@ -82,6 +82,11 @@ import { contentWillChange } from './listeners';
  * - delete last atom in a subsup: delete the subsup
  * - etc...
  *
+ * Note that `onDelete` may be called twice: once on the atom being deleted
+ * directly (in this case `branch` is `undefined`) and a second time
+ * on the parent if the atom was the first/last (in which case `branch` indicate
+ * which branch the atom was in).
+ *
  *
  * @param branch: if deleting inside an atom, the branch being deleted
  * (always the first or last atom of the branch). If undefined, the atom
@@ -97,29 +102,10 @@ function onDelete(
 ): boolean {
   const parent = atom.parent;
 
-  if (parent?.type === 'genfrac') {
-    let pos = model.offsetOf(atom.leftSibling);
-    parent.removeChild(atom);
-
-    if (parent.hasEmptyBranch('above') && parent.hasEmptyBranch('below')) {
-      // The last numerator or denominator of a fraction has been deleted:
-      // delete the fraction
-
-      pos = model.offsetOf(parent.leftSibling);
-      parent.parent!.removeChild(parent);
-      model.announce('delete', undefined, [parent]);
-      model.position = pos;
-      return true;
-    }
-    model.announce('delete', undefined, [atom]);
-    model.position = pos;
-    return true;
-  }
-
+  //
+  // 'leftright': \left\right
+  //
   if (parent && atom instanceof LeftRightAtom) {
-    //
-    // 'leftright': \left\right
-    //
     const atStart =
       (!branch && direction === 'forward') ||
       (branch === 'body' && direction === 'backward');
@@ -155,10 +141,10 @@ function onDelete(
     return true;
   }
 
+  //
+  // 'surd': square root
+  //
   if (parent && atom.type === 'surd') {
-    //
-    // 'surd': square root
-    //
     if (
       (direction === 'forward' && !branch) ||
       (direction === 'backward' && branch === 'body')
@@ -195,10 +181,10 @@ function onDelete(
     return true;
   }
 
+  //
+  // 'box': \boxed, \fbox 'enclose': \cancel
+  //
   if (parent && (atom.type === 'box' || atom.type === 'enclose')) {
-    //
-    // 'box': \boxed, \fbox 'enclose': \cancel
-    //
     const pos =
       (branch && direction === 'backward') ||
       (!branch && direction === 'forward')
@@ -213,6 +199,7 @@ function onDelete(
   //
   // 'genfrac': \frac, \choose, etc...
   //
+
   if (atom.type === 'genfrac' || atom.type === 'overunder') {
     if (!branch) {
       // After or before atom
@@ -253,14 +240,10 @@ function onDelete(
       return true;
     }
 
-    if (direction === 'backward') {
-      // Above first: move to before
+    if (direction === 'backward')
       model.position = model.offsetOf(atom.leftSibling);
-      return true;
-    }
+    else model.position = model.offsetOf(atom);
 
-    // Below last: move to after
-    model.position = model.offsetOf(atom);
     return true;
   }
 
@@ -321,6 +304,23 @@ function onDelete(
     return true;
   }
 
+  if (parent?.type === 'genfrac' && !branch && atom.type !== 'first') {
+    let pos = model.offsetOf(atom.leftSibling);
+    parent.removeChild(atom);
+    if (parent.hasEmptyBranch('above') && parent.hasEmptyBranch('below')) {
+      // The last numerator or denominator of a fraction has been deleted:
+      // delete the fraction
+      pos = model.offsetOf(parent.leftSibling);
+      parent.parent!.removeChild(parent);
+      model.announce('delete', undefined, [parent]);
+      model.position = pos;
+      return true;
+    }
+    model.announce('delete', undefined, [atom]);
+    model.position = pos;
+    return true;
+  }
+
   return false;
 }
 
@@ -374,7 +374,7 @@ export function deleteForward(model: ModelPrivate): boolean {
   return model.deferNotifications(
     { content: true, selection: true, type: 'deleteContentForward' },
     () => {
-      let target: Atom | null = model.at(model.position).rightSibling;
+      let target: Atom | undefined = model.at(model.position).rightSibling;
 
       if (target && onDelete(model, 'forward', target)) return;
 
@@ -386,7 +386,7 @@ export function deleteForward(model: ModelPrivate): boolean {
         )
           return;
 
-        target = null;
+        target = undefined;
       } else if (
         model.at(model.position).isLastSibling &&
         onDelete(model, 'forward', target.parent!, target.treeBranch)
@@ -445,6 +445,32 @@ export function deleteRange(
           model.offsetOf(parent.rightSibling),
         ];
       }
+    }
+
+    // If we have a placeholder denominator selected,
+    // hoist the denominator
+    if (
+      result.length === 1 &&
+      result[0].type === 'placeholder' &&
+      result[0].parent.type === 'genfrac'
+    ) {
+      const genfrac = result[0].parent!;
+      const branch = result[0].treeBranch === 'below' ? 'above' : 'below';
+      const pos = model.offsetOf(genfrac.leftSibling);
+      return model.deferNotifications(
+        { content: true, selection: true, type },
+        () => {
+          const numer = genfrac.removeBranch(branch);
+          if (!(numer.length === 1 && numer[0].type === 'placeholder')) {
+            const lastAtom = genfrac.parent!.addChildrenAfter(numer, genfrac);
+            genfrac.parent?.removeChild(genfrac);
+            model.position = model.offsetOf(lastAtom);
+          } else {
+            genfrac.parent?.removeChild(genfrac);
+            model.position = Math.max(0, pos);
+          }
+        }
+      );
     }
   }
   return model.deferNotifications(